+++ OgreMain/src/unzip.c
@@ -138,7 +138,7 @@ local int unzlocal_getByte(fin,pi)
     int *pi;
 {
     unsigned char c;
-    int err = fread(&c, 1, 1, fin);
+    size_t err = fread(&c, 1, 1, fin);
     if (err==1)
     {
         *pi = (int)c;
